feat: Created GetMatchedMoviesUseCase
This commit introduces a new use case, `GetMatchedMoviesUseCase`, responsible for fetching a list of movies that match specified criteria such as genre, runtime, and release date. It utilizes `MovieRepository` to retrieve data and `MovieMapper` to transform it into the domain model `Media`. Error handling is included to return an empty list in case of exceptions.